+++ appinfo/routes.php
@@ -13,6 +13,11 @@ return [
             'verb' => 'GET',
         ],
         [
+            'name' => 'Feed#cover',
+            'url' => '/feed/{token}/cover/{fileId}',
+            'verb' => 'GET',
+        ],
+        [
             'name' => 'Api#index',
             'url' => '/api/feeds',
             'verb' => 'GET',
@@ -28,9 +33,24 @@ return [
             'verb' => 'PUT',
         ],
         [
+            'name' => 'Api#update',
+            'url' => '/api/feeds/{id}',
+            'verb' => 'PUT',
+        ],
+        [
+            'name' => 'Api#uploadLogo',
+            'url' => '/api/feeds/{id}/logo',
+            'verb' => 'POST',
+        ],
+        [
             'name' => 'Api#destroy',
             'url' => '/api/feeds/{id}',
             'verb' => 'DELETE',
+        ],
+        [
+            'name' => 'Feed#logo',
+            'url' => '/feed/{token}/logo',
+            'verb' => 'GET',
         ],
     ],
 ];